requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.12.0=pyhd8ed1ab_0
aiohttp=3.7.4=py38h294d835_0
async-timeout=3.0.1=py_1000
attrs=20.3.0=pyhd3deb0d_0
blas=1.0=mkl
blinker=1.4=py_1
bokeh=2.3.1=py38haa244fe_0
bottleneck=1.3.2=py38h347fdf6_3
brotlipy=0.7.0=py38h294d835_1001
bzip2=1.0.8=he774522_0
ca-certificates=2021.4.13=haa95532_1
cachetools=4.2.1=pyhd8ed1ab_0
certifi=2020.12.5=py38haa244fe_1
cffi=1.14.5=py38hd8c33c5_0
cftime=1.4.1=py38h347fdf6_0
chardet=4.0.0=py38haa244fe_1
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
cryptography=3.4.7=py38hd7da0ea_0
cudatoolkit=10.2.89=h74a9793_1
curl=7.76.1=hf1763fc_1
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h294d835_3
dask=2021.4.0=pyhd8ed1ab_0
dask-core=2021.4.0=pyhd8ed1ab_0
distributed=2021.4.0=py38haa244fe_0
freetype=2.10.4=hd328e21_0
fsspec=2021.4.0=pyhd8ed1ab_0
future=0.18.2=py38haa244fe_3
google-auth=1.28.0=pyh44b312d_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.37.0=py38he5377a8_0
hdf4=4.2.13=h712560f_2
hdf5=1.10.4=h7ebc959_0
heapdict=1.0.1=py_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.10.1=py38haa244fe_0
intel-openmp=2020.2=254
jinja2=2.11.3=pyh44b312d_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=py38hd77b12b_0
krb5=1.17.2=hbae68bd_0
libcurl=7.76.1=hf1763fc_1
libnetcdf=4.7.3=h1302dcc_0
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.15.8=h7755175_0
libssh2=1.9.0=h680486a_6
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
locket=0.2.0=py_2
lz4-c=1.9.3=h2bbff1b_0
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=1.1.1=py38h294d835_3
matplotlib=3.3.4=py38haa95532_0
matplotlib-base=3.3.4=py38h49ac443_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.3.0=py38h46781fe_0
mkl_random=1.1.1=py38h47e9c7a_0
msgpack-python=1.0.2=py38hbd9d945_1
multidict=5.1.0=py38h294d835_1
netcdf4=1.5.3=py38h012c1a0_0
ninja=1.10.2=h6d14046_1
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.0.1=py_0
olefile=0.46=py_0
opencv=4.0.1=py38h2a7c758_0
openssl=1.1.1k=h8ffe710_0
packaging=20.9=pyh44b312d_0
pandas=1.2.4=py38h60cbd38_0
partd=1.2.0=pyhd8ed1ab_0
pillow=8.2.0=py38h4fa10fc_0
pip=21.0.1=py38haa95532_0
protobuf=3.15.8=py38h885f38d_0
psutil=5.8.0=py38h294d835_1
py-opencv=4.0.1=py38he44ac1e_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=2.0.1=pyhd8ed1ab_1
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38ha925a31_4
pysocks=1.7.1=py38haa244fe_3
python=3.8.8=hdbf39b2_5
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.8.1=py3.8_cuda10.2_cudnn7_0
pytorch-lightning=1.2.8=pyhd8ed1ab_0
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.4.1=py38h294d835_0
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
setuptools=52.0.0=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
sortedcontainers=2.3.0=pyhd8ed1ab_0
sqlite=3.35.4=h2bbff1b_0
tblib=1.7.0=pyhd8ed1ab_0
tensorboard=2.4.1=pyhd8ed1ab_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tk=8.6.10=he774522_0
toolz=0.11.1=py_0
torchaudio=0.8.1=py38
torchmetrics=0.2.0=pyhd8ed1ab_0
torchvision=0.9.1=py38_cu102
tornado=6.1=py38h294d835_1
tqdm=4.60.0=pyhd8ed1ab_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
xarray=0.17.0=pyhd8ed1ab_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yarl=1.6.3=py38h294d835_1
zict=2.0.0=py_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
